remove the `kernel.py` and move the `get_zoom_extends` to `utils.py` module
